+++ src/main/java/web/role/user/DeleteRoleUserServlet.java
@@ -27,15 +27,19 @@ public class DeleteRoleUserServlet extends HttpServlet {
         String idString = request.getParameter("id");
         User user = userService.selectById(Integer.parseInt(idString));
         String[] roleId = request.getParameterValues("roleId");
-        int deleteRoleUserSize = 0;
-        for (String s : roleId) {
-            if (!roleUserService.deleteRoleUser(Integer.parseInt(s),user.getId()).equals(0)) {
-                ++deleteRoleUserSize;
+        if (roleId != null) {
+            int deleteRoleUserSize = 0;
+            for (String s : roleId) {
+                if (!roleUserService.deleteRoleUser(Integer.parseInt(s),user.getId()).equals(0)) {
+                    ++deleteRoleUserSize;
+                }
             }
+            request.setAttribute("message","成功移除"+user.getNickname()+deleteRoleUserSize+"个角色!");
+        }else {
+            request.setAttribute("message","该用户没有角色!");
         }
-        request.setAttribute("message","成功移除"+user.getNickname()+deleteRoleUserSize+"个角色!");
+
         request.setAttribute("UserControlPage",true);
-        request.getSession().setAttribute("deleteRoleUser",null);
         request.getRequestDispatcher("pages/control.jsp").forward(request,response);
 
     }
